[Servo] Fixed reset when getting timer from PWM class
The problem was due to returning a copy of type GeneralPurposeTimer, in its constructor the timer is reset

+++ src/shared/actuators/Servo/ServoData.h
@@ -38,8 +38,8 @@ struct ServoData
 
     void print(std::ostream& os) const
     {
-        os << timestamp << "," << timer << "," << channel << "," << position
-           << "\n";
+        os << timestamp << "," << static_cast<int>(timer) << ","
+           << static_cast<int>(channel) << "," << position << "\n";
     }
 };
